third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/xmlhttprequest-sync-no-progress-events-expected.txt

CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/xmlhttprequest-sync-vs-async-assertion-failure.html' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
Test for:

bug 40996: Progress event should not be fired during synchronous XMLHttpRequest;
bug 17502: Assertion failure when trying to restart a sync XMLHttpRequest as an async one from onreadystatechange.
Step 1: Same origin request
readystatechange 4
load
loadend

Step 2: Cross origin request, disallowed
NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://localhost:8000/xmlhttprequest/xmlhttprequest-sync-vs-async-assertion-failure.html'.

Step 3: Cross origin request, allowed
readystatechange 4
load
loadend

Step 4: Cross origin request, check that preflight isn't attempted
readystatechange 4
load
loadend
